Update F - Equation.cpp
Given two numbers X and N. Print the result of the following equation:
S = (X0 - 1) + (X2) + (X4) + (X6) + ............... + (XN)

Note: Solve this problem using function and don't use built-in function.

Input
Only one line contains two numbers X and N (0 ≤ X, N ≤ 10)

Output
Print the equation result S.

Example
Input :
5 5
Output :
650

Note
First Test :
(50 - 1) + 52 + 54 = 0 + 25 + 625 = 650.

+++ Codeforces/ICPC SCU Newcomers Sheets '22/Week #5 (Functions)/F - Equation.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 
 // x^0 = 1 --> so , x^0 - 1 =0
-int Equation(int x,int n)
+long long Equation(int x,int n)
 {
-    int sum=0;
+    long long sum=0;
     for(int i=1;i<=n;i++)
     {
         if(i%2==0) sum+=pow(x,i);
